# [thrift-remodel] Add custom `PageLocation` decoder to speed up decoding of page indexes