@etseidl 2025-08-20
From #8160 benchmark results (old is the original thrift, new is the reworked decoder)
```
old
open(page index)        time:   [1.7446 ms 1.7524 ms 1.7639 ms]

new
open(page index)        time:   [698.00 µs 699.75 µs 701.66 µs]
```

With these changes the reading of the page index is now:
```
open(page index)        time:   [613.47 µs 614.98 µs 616.68 µs]
```

Note that the decoding of the column index is ~2/3 of the decoding time, so the speed up in the offset index decoding is considerable.

Is this worth the added complexity is the question. No changes to the `PageLocation` struct are on the horizon, so maybe.