@alamb 2025-08-26
> .now the benchmark is down to 584us.

Given the original description, does that mean that the orignal time was  1.75 ms  (thus this PR is like a 3X improvement)?

> open(page index)        time:   [1.7446 ms 1.7524 ms 1.7639 ms]